server: attempt to use some real gcode
There are some things I'm not familiar enough to address right now.
Backing away from this process for now. The internal axes definitions
are a little lost on me. They seem to position the head around a give
point. This breaks the internal kinematic solver if you attempt to keep
these values the same and change the X, Y, or Z positions.

Signed-off-by: Randolph Sapp <rs@ti.com>